+++ Tacta.EventStore/Projector/IProjection.cs
@@ -10,7 +10,7 @@ namespace Tacta.EventStore.Projector
 
         Task Initialize();
 
-        int GetSequence();
+        long GetSequence();
 
         Task Rebuild();
     }
