+++ Tacta.EventStore/Repository/IProjectionRepository.cs
@@ -4,7 +4,7 @@ namespace Tacta.EventStore.Repository
 {
     public interface IProjectionRepository
     {
-        Task<int> GetSequenceAsync();
+        Task<long> GetSequenceAsync();
         Task DeleteAllAsync();
     }
 }
